% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_analysis.R
\name{active_rels}
\alias{active_rels}
\title{Compile Active Relationships}
\usage{
active_rels(x)
}
\arguments{
\item{x}{is the netsim output}
}
\description{
This function takes the output from simulation cel.temp and splits
the data by network type
}
