% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_analysis.R
\name{as_nsfg_rels}
\alias{as_nsfg_rels}
\title{Sample Relationships with NSFG Design}
\usage{
as_nsfg_rels(x)
}
\arguments{
\item{x}{is the netsim output}
}
\description{
This function takes the output from simulation cel.temp and cel.complete,
removes relationships that ended more than 12 months from last time step,
adds exact duration flag and limits to 3 relationships per ID.
}
