% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_analysis.R
\name{calc_meandeg}
\alias{calc_meandeg}
\title{Final Mean Degree Distribution of Simulation}
\usage{
calc_meandeg(x, maxage)
}
\arguments{
\item{x}{network simulation object}

\item{maxage}{age at which egos depart model - either 45 or 65}
}
\description{
This function calculates the target mean degree and the mean degree at the end of
the simulation (averaged across each run) for each network.
}
