% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_analysis.R
\name{complete_rels}
\alias{complete_rels}
\title{Compile Active Relationships}
\usage{
complete_rels(x)
}
\arguments{
\item{x}{is the netsim output}
}
\description{
This function takes the output from simulation cel.complete,
keeps all relationships that begin after the simulation began, and splits
the data by network type
}
