% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_analysis.R
\name{duration_calcs}
\alias{duration_calcs}
\title{Calculate Active and Completed Mean Relationship Length}
\usage{
duration_calcs(x, maxage)
}
\arguments{
\item{x}{network simulation object}

\item{maxage}{age at which egos depart model - either 45 or 65}
}
\description{
This function calculates the mean relationship length of active rels (averaged
across each run) in each network, the mean length of completed rels, and includes
the target length in the exported object.
}
