% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egodata_meandegs.R
\name{egodata_meandegs}
\alias{egodata_meandegs}
\title{Mean Degree Distributions in Egodata}
\usage{
egodata_meandegs(categorical = FALSE)
}
\description{
This function takes the egodata (stored in system data) and summarizes the expected mean degree
distribution by age.
}
\examples{
egodata_meandegs()
}
