% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_analysis.R
\name{get_agedist}
\alias{get_agedist}
\title{Get Active Agedist}
\usage{
get_agedist(x, categorical = FALSE)
}
\arguments{
\item{x}{is the netsim output}
}
\description{
This function takes final age attribute and generates a dist of counts
by age (only 15:44 counts), averaged across all simulation runs
}
