% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_small_netsim.R
\name{make_small_netsim}
\alias{make_small_netsim}
\title{Netsim Size Reduction}
\usage{
make_small_netsim(x)
}
\arguments{
\item{x}{is the netsim output}
}
\description{
Take objects and remove the cel.complete and misc variables I don't currently need
}
