% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_analysis.R
\name{marcoh_dist}
\alias{marcoh_dist}
\title{Distribution of Mean Degree by Age in Marriage/Cohab Network}
\usage{
marcoh_dist(
  x,
  maxage,
  marcoh_ages_restricted,
  marcoh_ages_all,
  categorical = FALSE
)
}
\arguments{
\item{x}{network simulation object}

\item{maxage}{age at which egos depart model - either 45 or 65}

\item{marcoh_ages_restricted}{dist of expected mean degree using age-restricted alter set}

\item{marcoh_ages_all}{dist of exp mean degree using all reported alters}

\item{categorical}{default = FALSE, whether or not to return dist by age or agecat}
}
\description{
This function takes the list of relationships active on the last time step of
the simulation (averaged over each run) and calculates the mean degree by age. It
then takes that distribution and plots it under the expected distribution based on
the egodata. The exported object includes both the ggplot object and the raw data.
}
